2023.03.22

+++ src/main/java/com/hhs/codeboard/blog/jpa/entity/board/dto/request/BoardArticleRequest.java
@@ -10,10 +10,8 @@ import java.time.LocalDate;
 @Setter
 public class BoardArticleRequest {
 
-    
-
     @Parameter
-    private Integer seq;
+    private Long seq;
     private String title;
     private String content;
     private String summary;
